refactor(web): extract SelectableOption into shared/ui
